+++ pom.xml
@@ -557,6 +557,7 @@
 
         <!-- Languages -->
         <module>querydsl-scala</module>
+        <module>querydsl-kotlin</module>
       </modules>
     </profile>
