+++ querydsl-codegen/src/test/java/com/querydsl/codegen/EmbeddableSerializerTest.java
@@ -35,7 +35,7 @@ public class EmbeddableSerializerTest {
 
     private final TypeMappings typeMappings = new JavaTypeMappings();
 
-    private final EntitySerializer serializer = new EmbeddableSerializer(typeMappings, Collections.<String>emptySet());
+    private final EntitySerializer serializer = new DefaultEmbeddableSerializer(typeMappings, Collections.<String>emptySet());
 
     private final StringWriter writer = new StringWriter();
