+++ querydsl-codegen/src/test/java/com/querydsl/codegen/EntitySerializerTest.java
@@ -35,7 +35,7 @@ public class EntitySerializerTest {
 
     private final TypeMappings typeMappings = new JavaTypeMappings();
 
-    private final EntitySerializer serializer = new EntitySerializer(typeMappings, Collections.<String>emptySet());
+    private final EntitySerializer serializer = new DefaultEntitySerializer(typeMappings, Collections.<String>emptySet());
 
     private final StringWriter writer = new StringWriter();
