+++ querydsl-codegen/src/test/java/com/querydsl/codegen/ProjectionSerializerTest.java
@@ -40,7 +40,7 @@ public class ProjectionSerializerTest {
         type.addConstructor(new Constructor(Arrays.asList(firstName, lastName, age)));
 
         Writer writer = new StringWriter();
-        ProjectionSerializer serializer = new ProjectionSerializer(new JavaTypeMappings());
+        ProjectionSerializer serializer = new DefaultProjectionSerializer(new JavaTypeMappings());
         serializer.serialize(type, SimpleSerializerConfig.DEFAULT, new JavaWriter(writer));
         assertTrue(writer.toString().contains("Expression<String> firstName"));
         assertTrue(writer.toString().contains("Expression<String> lastName"));
