+++ querydsl-kotlin/src/main/kotlin/com/querydsl/kotlin/KotlinSuperSerializer.kt
@@ -1,0 +1,9 @@
+package com.querydsl.kotlin
+
+import com.querydsl.codegen.SupertypeSerializer
+import com.querydsl.codegen.TypeMappings
+import javax.inject.Inject
+import javax.inject.Named
+
+class KotlinSuperSerializer @Inject constructor(mappings: TypeMappings, @Named("keywords") keyword: Collection<String>) : KotlinEntitySerializer(mappings, keyword), SupertypeSerializer {
+}
