+++ querydsl-kotlin/src/main/kotlin/com/querydsl/kotlin/KotlinTypeMappings.kt
@@ -1,0 +1,5 @@
+package com.querydsl.kotlin
+
+import com.querydsl.codegen.JavaTypeMappings
+
+class KotlinTypeMappings : JavaTypeMappings()
