+++ querydsl-kotlin/src/main/resources/META-INF/services/com.querydsl.apt.Extension
@@ -1,0 +1,1 @@
+com.querydsl.kotlin.KotlinAptExtension
